keyboard and controller support on character selection menu
Just added controller and keyboard support to the character selection menu. A network simulator window pops up and currently the only way to get rid of it is to use your mouse, but if you do not touch it, everything still works fine. Will make a fix for that tomorrow.